modules/autoscale/src/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client==2.86.0
pan-os-python
google-cloud-secret-manager